requirements.txt
motor==3.3.2
pymongo==4.6.0
pydantic==2.5.0
python-dotenv
pyrofork
TgCrypto-pyrofork
aiohttp